сделал регистрацию города

+++ src/main/java/ru/nsu/concertsmate/backend/services/impl/UserCitiesServiceImpl.java
@@ -7,11 +7,17 @@ import org.springframework.stereotype.Service;
 import ru.nsu.concertsmate.backend.model.dto.UserCityDto;
 import ru.nsu.concertsmate.backend.model.entities.User;
 import ru.nsu.concertsmate.backend.model.entities.UserCity;
+import ru.nsu.concertsmate.backend.model.entities.UserCityEmbedded;
 import ru.nsu.concertsmate.backend.repositories.UsersCitiesRepository;
 import ru.nsu.concertsmate.backend.repositories.UsersRepository;
 import ru.nsu.concertsmate.backend.services.UsersCitiesService;
+import ru.nsu.concertsmate.backend.services.exceptions.CityAlreadyExistsException;
+import ru.nsu.concertsmate.backend.services.exceptions.NoSuchCityException;
+import ru.nsu.concertsmate.backend.services.exceptions.NoSuchUserException;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.logging.SocketHandler;
 
 @Service
 public class UserCitiesServiceImpl implements UsersCitiesService {
@@ -28,26 +34,46 @@ public class UserCitiesServiceImpl implements UsersCitiesService {
 
 
     @Override
-    public UserCityDto saveUserCity(long telegramId, String cityName) {
-        long userId = usersRepository.findByTelegramId(telegramId).get().getId();
-        UserCity userCity = new UserCity(userId, cityName);
-        userCity = usersCitiesRepository.save(userCity);
+    public UserCityDto saveUserCity(long telegramId, String cityName) throws NoSuchUserException, CityAlreadyExistsException {
+        Optional<User> user = usersRepository.findByTelegramId(telegramId);
+        if (user.isEmpty()){
+            throw new NoSuchUserException();
+        }
+        Optional<UserCity> testUnique = usersCitiesRepository.findById(new UserCityEmbedded(user.get().getId(), cityName));
+        if (testUnique.isPresent()){
+            throw new CityAlreadyExistsException();
+        }
+        UserCity userCity = new UserCity(user.get().getId(), cityName);
+        UserCity result = usersCitiesRepository.save(userCity);
+        if (!result.equals(userCity)){
+            throw new RuntimeException("can't save user city");
+        }
 
-        return modelMapper.map(userCity, UserCityDto.class);
+        return modelMapper.map(result, UserCityDto.class);
     }
 
     @Override
-    public UserCityDto deleteUserCity(long telegramId, String cityName) {
-        long userId = usersRepository.findByTelegramId(telegramId).get().getId();
-        UserCity userCity = new UserCity(userId, cityName);
+    public UserCityDto deleteUserCity(long telegramId, String cityName) throws NoSuchUserException, NoSuchCityException {
+        Optional<User> user = usersRepository.findByTelegramId(telegramId);
+        if (user.isEmpty()){
+            throw new NoSuchUserException();
+        }
+        Optional<UserCity> testUnique = usersCitiesRepository.findById(new UserCityEmbedded(user.get().getId(), cityName));
+        if (testUnique.isEmpty()){
+            throw new NoSuchCityException();
+        }
+        UserCity userCity = new UserCity(user.get().getId(), cityName);
         usersCitiesRepository.delete(userCity);
 
         return null;
     }
 
     @Override
-    public List<String> getUserCities(long telegramId) {
-        long userId = usersRepository.findByTelegramId(telegramId).get().getId();
-        return usersCitiesRepository.getUserCities(userId).get();
+    public List<String> getUserCities(long telegramId) throws NoSuchUserException {
+        Optional<User> user = usersRepository.findByTelegramId(telegramId);
+        if (user.isEmpty()){
+            throw new NoSuchUserException();
+        }
+        return usersCitiesRepository.getUserCities(user.get().getId()).get();
     }
 }
